Moved EnvChange to codec package
